fix: error handling when internet is unavailable
It could also be triggered in case the domain name couldnt be resolved
or the connection to server fails for any other reason.
Since there is no way to reliably determine to cause, it exits the
program with a generic connection error and then prints the original
error message for the user to diagnose.